TEST: [cmn][cli] Promote client testing utils to common testutils mod
